refact: move Gateway to appropriate package for better module organization
