+++ src/main/kotlin/com/dobby/backend/application/service/MemberService.kt
@@ -4,7 +4,7 @@ import com.dobby.backend.application.usecase.member.CreateParticipantUseCase
 import com.dobby.backend.application.usecase.member.CreateResearcherUseCase
 import com.dobby.backend.application.usecase.member.VerifyResearcherEmailUseCase
 import com.dobby.backend.domain.exception.SignupOauthEmailDuplicateException
-import com.dobby.backend.domain.gateway.MemberGateway
+import com.dobby.backend.domain.gateway.member.MemberGateway
 import com.dobby.backend.infrastructure.database.entity.enum.MemberStatus
 import jakarta.transaction.Transactional
 import org.springframework.stereotype.Service
