+++ src/main/kotlin/com/dobby/backend/application/usecase/auth/GenerateTokenWithRefreshTokenUseCase.kt
@@ -1,8 +1,8 @@
 package com.dobby.backend.application.usecase.auth
 
 import com.dobby.backend.application.usecase.UseCase
-import com.dobby.backend.domain.gateway.MemberGateway
-import com.dobby.backend.domain.gateway.TokenGateway
+import com.dobby.backend.domain.gateway.member.MemberGateway
+import com.dobby.backend.domain.gateway.auth.TokenGateway
 import com.dobby.backend.domain.model.member.Member
 
 class GenerateTokenWithRefreshTokenUseCase(
