+++ src/main/kotlin/com/dobby/backend/application/usecase/experiment/GetExperimentPostCountsByRegionUseCase.kt
@@ -1,7 +1,7 @@
 package com.dobby.backend.application.usecase.experiment
 
 import com.dobby.backend.application.usecase.UseCase
-import com.dobby.backend.domain.gateway.ExperimentPostGateway
+import com.dobby.backend.domain.gateway.experiment.ExperimentPostGateway
 import com.dobby.backend.infrastructure.database.entity.enum.areaInfo.Region
 
 class GetExperimentPostCountsByRegionUseCase(
