+++ src/main/kotlin/com/dobby/backend/application/usecase/member/VerifyResearcherEmailUseCase.kt
@@ -4,8 +4,7 @@ package com.dobby.backend.application.usecase.member
 import com.dobby.backend.application.usecase.UseCase
 import com.dobby.backend.domain.exception.EmailNotValidateException
 import com.dobby.backend.domain.exception.VerifyInfoNotFoundException
-import com.dobby.backend.domain.gateway.VerificationGateway
-import com.dobby.backend.domain.model.Verification
+import com.dobby.backend.domain.gateway.email.VerificationGateway
 import com.dobby.backend.infrastructure.database.entity.enum.VerificationStatus
 
 class VerifyResearcherEmailUseCase(
