+++ src/main/kotlin/com/dobby/backend/domain/gateway/auth/NaverAuthGateway.kt
@@ -1,4 +1,4 @@
-package com.dobby.backend.domain.gateway.feign
+package com.dobby.backend.domain.gateway.auth
 
 import com.dobby.backend.presentation.api.dto.response.auth.naver.NaverInfoResponse
 import com.dobby.backend.presentation.api.dto.response.auth.naver.NaverTokenResponse
