+++ src/main/kotlin/com/dobby/backend/domain/gateway/email/EmailGateway.kt
@@ -1,4 +1,4 @@
-package com.dobby.backend.domain.gateway
+package com.dobby.backend.domain.gateway.email
 
 interface EmailGateway {
     fun sendEmail(to: String, subject: String, content: String)
