+++ src/main/kotlin/com/dobby/backend/domain/gateway/email/VerificationGateway.kt
@@ -1,4 +1,4 @@
-package com.dobby.backend.domain.gateway
+package com.dobby.backend.domain.gateway.email
 
 import com.dobby.backend.domain.model.Verification
 import com.dobby.backend.infrastructure.database.entity.enum.VerificationStatus
