+++ src/main/kotlin/com/dobby/backend/domain/gateway/member/ResearcherGateway.kt
@@ -1,4 +1,4 @@
-package com.dobby.backend.domain.gateway
+package com.dobby.backend.domain.gateway.member
 
 import com.dobby.backend.domain.model.member.Member
 import com.dobby.backend.domain.model.member.Researcher
