+++ src/main/kotlin/com/dobby/backend/infrastructure/gateway/member/ResearcherGatewayImpl.kt
@@ -1,9 +1,7 @@
-package com.dobby.backend.infrastructure.gateway
+package com.dobby.backend.infrastructure.gateway.member
 
-import com.dobby.backend.domain.gateway.ResearcherGateway
-import com.dobby.backend.domain.model.member.Member
+import com.dobby.backend.domain.gateway.member.ResearcherGateway
 import com.dobby.backend.domain.model.member.Researcher
-import com.dobby.backend.infrastructure.converter.MemberConverter
 import com.dobby.backend.infrastructure.converter.ResearcherConverter
 import com.dobby.backend.infrastructure.database.repository.ResearcherRepository
 import org.springframework.stereotype.Component
@@ -11,7 +9,7 @@ import org.springframework.stereotype.Component
 @Component
 class ResearcherGatewayImpl(
     private val researcherRepository: ResearcherRepository
-) : ResearcherGateway{
+) : ResearcherGateway {
 
     override fun findByMemberId(memberId: Long) : Researcher? {
         val researcherEntity = researcherRepository
