+++ src/test/kotlin/com/dobby/backend/application/usecase/auth/GenerateTestTokenUseCaseTest.kt
@@ -1,9 +1,8 @@
 package com.dobby.backend.application.usecase.auth
 
-import com.dobby.backend.application.usecase.auth.GenerateTestTokenUseCase
-import com.dobby.backend.domain.gateway.MemberGateway
+import com.dobby.backend.domain.gateway.member.MemberGateway
 import io.kotest.core.spec.style.BehaviorSpec
-import com.dobby.backend.domain.gateway.TokenGateway
+import com.dobby.backend.domain.gateway.auth.TokenGateway
 import com.dobby.backend.domain.model.member.Member
 import com.dobby.backend.infrastructure.database.entity.enum.MemberStatus
 import com.dobby.backend.infrastructure.database.entity.enum.ProviderType
